+++ backend/src/neuroagent/tools/bluenaas_scs_post.py
@@ -24,7 +24,7 @@ logger = logging.getLogger(__name__)
 class SCSPostMetadata(BaseMetadata):
     """Metadata class for the get all simulations api."""
 
-    token: str
+    httpx_client: AsyncClient
     vlab_id: str
     project_id: str
     bluenaas_url: str
@@ -148,14 +148,9 @@ class SCSPostTool(BaseTool):
 
         response = await self.metadata.httpx_client.post(
             url=f"{self.metadata.bluenaas_url}/simulation/single-neuron/{params.virtual_lab_id}/{params.project_id}/run",
-            params={
-                k: v
-                for k, v in params.model_dump(
-                    exclude={"virtual_lab_id", "project_id", "body"}
-                ).items()
-                if v is not None
-            },
-            headers={"Authorization": f"Bearer {self.metadata.token}"},
+            params=params.model_dump(
+                exclude_defaults=True, exclude={"virtual_lab_id", "project_id", "body"}
+            ),
             json=params.body.model_dump(),
         )
         json_response = response.json()